+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Fixed
 
 - Wrong handling of command-line arguments in `east util connect`.
+- Suppress `integer to pointer cast pessimizes optimization opportunities`, caused by LOG_* macros, 
+  when `east codechecker check` is run.
 
 ## [0.25.1] - 2024-12-18